feat(onboarding): slim /start to two steps (school type → identity)
Replace the lengthy multi-step /start wizard with a two-step flow:
① School type (Basic / Senior + SHS·SHTS track / Multi-tier)
② School identity + owner details.

Everything else (calendar, classes, subjects, grade scale, fees) is
auto-seeded from tier defaults by onboardSchool; the owner finishes the
rest from the post-signup §01 checklist or via super-admin setup.

- Pricing CTAs pre-select the type (/start?type=BASIC|SENIOR); when set,
  the wizard skips step ① straight to identity (a "Change" link reveals
  it). Bare /start does both steps.
- Owner → ADMIN: the identity step captures the person who signs in
  first. Headmaster is now optional (created only when name+phone given),
  collected later or by super-admin.
- Preserve the full multi-step wizard at components/onboarding/full-wizard.tsx
  (FullOnboardingWizard) for reuse in the super-admin tenant-setup portal;
  not routed anywhere in the app.
- Dashboard §01 guide: reword step 1 to "Review your school setup" (a new
  school comes seeded), add the "What to expect" timeline and the
  "Need help?" side panel per schoolup-empty-states §01.
- Remove dead headmasterPhone field from the onboardSchool return object.
