+++ omnischools/app/(marketing)/start/page.tsx
@@ -1,13 +1,27 @@
 import type { Metadata } from "next";
 import { OnboardingWizard } from "@/components/onboarding/wizard";
+import type { CardId } from "@/lib/onboarding";
 
 export const metadata: Metadata = {
   title: "Onboard your school",
   description:
-    "Set up your school on Omnischools — identity, school type, calendar, structure, staff and billing. Basic schools finish in six steps; SHS adds residency and WAEC. Your GES-standard calendar is configured automatically.",
+    "Set up your school on Omnischools in two quick steps — school type and school identity. Your GES-standard calendar, classes and grade scale are configured automatically; you finish the rest from a guided checklist after signing in.",
 };
 
-export default function StartPage() {
+const VALID_TYPES = ["BASIC", "SENIOR", "MULTI"] as const;
+
+export default function StartPage({
+  searchParams,
+}: {
+  searchParams: { type?: string };
+}) {
+  // A pricing plan may pre-select the school type (?type=BASIC|SENIOR); when it does,
+  // the wizard skips straight to School identity.
+  const raw = searchParams.type?.toUpperCase();
+  const initialType = (VALID_TYPES as readonly string[]).includes(raw ?? "")
+    ? (raw as CardId)
+    : undefined;
+
   return (
     <main className="mx-auto max-w-5xl px-4 py-12 sm:px-6">
       <div className="mb-7 text-center">
@@ -19,11 +33,13 @@ export default function StartPage() {
           <em className="not-italic text-gold [font-style:italic]">school.</em>
         </h1>
         <p className="mt-3 text-base text-navy-2">
-          Six short steps — Basic schools finish at six; SHS adds two. No card, no
-          commitment — a 30-day trial starts the moment you&apos;re in.
+          {initialType
+            ? "One step — tell us who you are. Your calendar, classes and grade scale come pre-configured; you finish setup from a guided checklist after signing in."
+            : "Two short steps — your school type and who you are. Everything else is pre-configured and completed from a guided checklist after you sign in."}{" "}
+          No card, no commitment — a 30-day trial starts the moment you&apos;re in.
         </p>
       </div>
-      <OnboardingWizard />
+      <OnboardingWizard initialType={initialType} />
     </main>
   );
 }
